+++ frontend/src/components/Homepage/index.js
@@ -60,6 +60,7 @@ function Homepage() {
                         value={editedPlanName}
                         onChange={(e) => setEditedPlanName(e.target.value)}
                         className="flex_col_center"
+                        maxLength={40}
                     />
                     <p onClick={() => setEditPlanNameVisibility(false)}>Cancel</p>
                 </div>
@@ -70,8 +71,11 @@ function Homepage() {
 
     // Delete the meal plan
     const deletePlan = async () => {
-        await dispatch(deleteMealPlan(selectedPlan.id));
-        dispatch(deleteSelectedPlan());
+        let confirm = window.confirm('This will permanently delete this meal plan')
+        if (confirm) {
+            await dispatch(deleteMealPlan(selectedPlan.id));
+            dispatch(deleteSelectedPlan());
+        }
     }
 
 
@@ -102,6 +106,7 @@ function Homepage() {
                     value={newDayName}
                     onChange={(e) => setNewDayName(e.target.value)}
                     className="daily_schedule_header flex_col_center"
+                    maxLength={40}
                 />
                 <p onClick={() => setNewDayFormVisibility(false)}>Cancel</p>
             </form>
